A number of changes needed to get this building as an x64 plug-in & appropriate changes to remove workarounds no longer needed Changed to use more methods from the wacup core to reduce OS imports, etc Added support to allow for custom preview & overlay icons from the current skin or from the user settings folder - will look in 'Taskbar' as needed for them before ending up back on the default icon resources Fix some WINE related issues if trying to load / use the plug-in Fix some odd crashes seen

+++ taskbar.cpp
@@ -47,10 +47,10 @@ HRESULT iTaskBar::ThumbBarUpdateButtons(std::vector<THUMBBUTTON>& buttons, HIMAG
 	if (ImageList != NULL)
 	{
 		pTBL->ThumbBarSetImageList(plugin.hwndParent, ImageList);
-		return pTBL->ThumbBarAddButtons(plugin.hwndParent, buttons.size(), &buttons[0]);
+		return pTBL->ThumbBarAddButtons(plugin.hwndParent, (int)buttons.size(), &buttons[0]);
 	}
 
-	return pTBL->ThumbBarUpdateButtons(plugin.hwndParent, buttons.size(), &buttons[0]);
+	return pTBL->ThumbBarUpdateButtons(plugin.hwndParent, (int)buttons.size(), &buttons[0]);
 }
 
 void iTaskBar::SetIconOverlay(HICON icon, const std::wstring &text)
